Finish Kalman filter integration. Bugs to be fixed.

+++ workspace/assignments/04-imu-lidar-gnss-fusion/src/lidar_localization/src/publisher/cloud_publisher.cpp
@@ -17,7 +17,7 @@ CloudPublisher::CloudPublisher(ros::NodeHandle& nh,
 }
 
 void CloudPublisher::Publish(CloudData::CLOUD_PTR&  cloud_ptr_input, double time) {
-    ros::Time ros_time((float)time);
+    ros::Time ros_time(time);
     PublishData(cloud_ptr_input, ros_time);
 }
