+++ workspace/assignments/04-imu-lidar-gnss-fusion/src/lidar_localization/src/publisher/tf_broadcaster.cpp
@@ -14,7 +14,7 @@ TFBroadCaster::TFBroadCaster(std::string frame_id, std::string child_frame_id) {
 
 void TFBroadCaster::SendTransform(Eigen::Matrix4f pose, double time) {
     Eigen::Quaternionf q(pose.block<3,3>(0,0));
-    ros::Time ros_time((float)time);
+    ros::Time ros_time(time);
     transform_.stamp_ = ros_time;
     transform_.setRotation(tf::Quaternion(q.x(), q.y(), q.z(), q.w()));
     transform_.setOrigin(tf::Vector3(pose(0,3), pose(1,3), pose(2,3)));
